% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_yearly}
\alias{aggregate_yearly}
\title{aggregate 8-day 3d array to yearly}
\usage{
aggregate_yearly(array, dates, type = c("sum", "mean")[1])
}
\description{
aggregate 8-day 3d array to yearly
}
