% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_3d.R
\name{apply_3d}
\alias{apply_3d}
\alias{apply_row}
\title{strengthened apply function for 3d array}
\usage{
apply_3d(array, dim = 3, FUN = rowMeans2, by = NULL, na.rm = TRUE, ...)

apply_row(mat, by, FUN = rowMeans2, w = NULL, na.rm = TRUE, ...)
}
\arguments{
\item{array}{3d or 2d array}

\item{dim}{The dim to apply function, if \code{dim > ndim(array)}, dim will be reset to
\code{ndim(array)}.}
}
\description{
strengthened apply function for 3d array
}
