% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{check_nc_att}
\alias{check_nc_att}
\title{Check that a list describes a netCDF attribute}
\usage{
check_nc_att(x)
}
\arguments{
\item{x}{The list to check.}
}
\description{
Ensures the list contains only 'name' (character), 'value' (atomic), and 'type' (in
\link{NC_TYPES})
}
\seealso{
Use \code{\link[=nc_att]{nc_att()}} to create a valid list describing a netCDF attribute
}
