% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{fill_seq}
\alias{fill_seq}
\title{Fill gaps in a sequence}
\usage{
fill_seq(
  x,
  by = NULL,
  tolerance = sqrt(.Machine$double.eps),
  decreasing = FALSE
)
}
\arguments{
\item{x}{Numeric vector whose values differ by integer multiples of an increment e.g. c(1, 2, 4).}

\item{by}{The increment. If \code{NULL}, \code{by} is set to the smallest difference between values.}

\item{tolerance}{Tolerance to allow for floating point errors and imprecision. Values that differ
by < \code{tolerance} are deduplicated. Differences must be within \code{tolerance} of an integer multiple
of \code{by}.}

\item{decreasing}{Whether to sort the values in decreasing order.}
}
\value{
A increasing unique numeric vector whose values differ by the increment.
}
\description{
Fill gaps in a sequence
}
\examples{
## Fill a sequence that contains gaps
fill_seq(c(1, 2, 3, 5, 7)) # -> c(1, 2, 3, 4, 5, 6, 7)

## Specify the increment
fill_seq(c(1, 3, 5), by = 1) # -> c(1, 2, 3, 4, 5)

## Raises if differences between values are not integer multiples of increment
\dontrun{
fill_seq(c(1, 2, 3, 5.001))
}

## Allow slight variation in the differences between values
fill_seq(c(1, 2, 3, 5.001), tolerance = 1e-2) # -> c(1.000, 2.000, 3.000, 4.000, 5.001)
}
