% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{get_names}
\alias{get_names}
\title{Iterate over a list of named lists, returning the 'name' element of each}
\usage{
get_names(x)
}
\arguments{
\item{x}{A list of named lists.}
}
\value{
A character vector of names, or zero-length character vector if x is not a list
}
\description{
Iterate over a list of named lists, returning the 'name' element of each
}
