% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{get_nctype}
\alias{get_nctype}
\title{Return the netCDF type corresponding to an object's class}
\usage{
get_nctype(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The name of a netCDF type.
}
\description{
Return the netCDF type corresponding to an object's class
}
