% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncread.R
\name{getCMIP5date}
\alias{getCMIP5date}
\title{getCMIP5date}
\usage{
getCMIP5date(nslice, file, calendar = "365", origin, check_date = FALSE)
}
\arguments{
\item{nslice}{length of time in original nc file.}

\item{file}{Character, nc file name.}

\item{calendar}{the calendar type.}

\item{origin}{a \code{Date} object, or something which can be coerced by
    \code{as.Date(origin, \dots)} to such an object.}

\item{check_date}{If true, check whether date in nc file is consistent with
\code{date_begin} and \code{date_end} in its filename.}
}
\value{
Date character.
}
\description{
For \verb{360_day} calendar, converting to \code{Date} will lead to error.
}
\examples{
\dontrun{
getCMIP5date(nslice, file, calendar = "365")
}
}
\keyword{internal}
