% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_date.R
\name{guess_date}
\alias{guess_date}
\title{guess CMIP5 date according to nc file name}
\usage{
guess_date(nslice, file, calendar = "365", origin, check_date = FALSE)
}
\arguments{
\item{nslice}{length of time in original nc file.}

\item{file}{Character, nc file name.}

\item{calendar}{the calendar type.}

\item{check_date}{If true, check whether date in nc file is consistent with
\code{date_begin} and \code{date_end} in its filename.}
}
\value{
PCICt date vector
}
\description{
guess CMIP5 date according to nc file name
}
\examples{
\dontrun{
guess_date(nslice, file, calendar = "365")
}
}
\keyword{internal}
