% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2_interp3dBilinear.R
\name{interp3d_bilinear}
\alias{interp3d_bilinear}
\title{Bilinear interpolation for 3d array}
\usage{
interp3d_bilinear(
  z,
  grid = NULL,
  grid_target,
  na.rm = TRUE,
  convertTo2d = FALSE
)
}
\arguments{
\item{z}{3d array (lon, lat, time), with the dimension of \verb{[nlon, nlat, ntime]}. \code{image(z[,,1])} should look normal.}

\item{grid}{A list object with lon\verb{, }lat\code{and}loc`.
\itemize{
\item \code{lon}: numeric, \verb{[nlon]}
\item \code{lat}: numeric, \verb{[nlon]}
}}

\item{grid_target}{A list object with \code{lon}, \code{lat} and \code{loc}.
\itemize{
\item \code{lon}: numeric, \verb{[nlon]}
\item \code{lat}: numeric, \verb{[nlon]}
\item \code{loc}: a data.frame with the column of c("x", "y"), with the row length of
\code{nlon*nlat}.
}}

\item{na.rm}{If true, Na value in margin will be fixed.}
}
\description{
Bilinear interpolation for 3d array
}
\note{
High resolution interpolating to low resolution will lead to unreliable
result, if \code{cellsize.new/cellsize.origin > 2}.
}
\examples{
\dontrun{
library(CMIP6tools)
# library(rgdal)
# library(sp2)
library(maptools)

# polyline
read_polyline <- function(file, lwd = 0.5, ...){
    suppressWarnings(shp <- readShapeLines(file))
    proj4string(shp) <- prj84
    list("sp.lines", shp, first = FALSE, lwd = lwd, ...) # china poly    
}
# sp_arc_world <- read_polyline("G:/ArcGIS/continent.shp") # sp_arc_world
# sp_arc_CH <- read_polyline(system.file("extdata/shp", "bou1_4l_sml.shp", package = "CMIP5tools"))

range <- c(70, 140, 15, 55)
# MAIN SCRIPTS ------------------------------------------------------------
x <- readRDS(system.file("extdata", "tasmax_Threshold_ACCESS1-0.RDS", package = "CMIP5tools"))

class(x) <- "raster2"
plot(x)

x_interp <- interp3d_bilinear(x$grid, x$TRS, range, cellsize_x = 1.5) 
x_clip <- clip_raster2(x, range)

plot(x_interp)
plot(x_clip) # , sp.layout = sp_arc_CH
}
}
