% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{log_attribute}
\alias{log_attribute}
\title{Log a netCDF attribute and its value}
\usage{
log_attribute(log_fun, varname, attribute)
}
\arguments{
\item{log_fun}{Logging function e.g. \code{\link[logger:log_level]{logger::log_debug()}}.}

\item{varname}{Name of the NetCDF variable that the attribute describes.}

\item{attribute}{NetCDF attribute defined with \code{\link[=nc_att]{nc_att()}}.}
}
\description{
Logs the name of a netCDF attribute and its value. The value is quoted if it is of type NC_CHAR.
}
