% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_tools.R
\name{nc_aggregateToYearly}
\alias{nc_aggregateToYearly}
\title{Aggregate to yearly scale}
\usage{
nc_aggregateToYearly(infile, outfile, vars_mean = NULL, overwrite = FALSE)
}
\arguments{
\item{infile, outfile}{INPUT and OUTPUT nc files}

\item{vars_mean}{variables which are aggregated by \code{mean}. The others will be
\code{sum}.}
}
\description{
Aggregate to yearly scale
}
\note{
Current function can't handle with NA vlaues.
}
