% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{nc_att}
\alias{nc_att}
\title{Define a netCDF attribute}
\usage{
nc_att(name, value, type = NULL)
}
\arguments{
\item{name}{Name of the attribute.}

\item{value}{Value of the attribute.}

\item{type}{NetCDF type for \code{value}. If \code{NULL}, \code{type} is determined from \code{value}'s class. See
\link{NC_TYPES} for a list of valid types.}
}
\value{
A list of the passed arguments.
}
\description{
Define a netCDF attribute
}
\examples{
nc_att("Test", value = "Test attribute", type = "NC_CHAR")
}
