% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{nc_coord}
\alias{nc_coord}
\title{Define a netCDF coordinate variable (dimension and variable)}
\usage{
nc_coord(
  name,
  values,
  type = NULL,
  long_name = NULL,
  standard_name = NULL,
  axis = NULL,
  units = NULL,
  atts = NULL,
  decreasing = FALSE,
  unlim = FALSE,
  fill = FALSE,
  ...
)
}
\arguments{
\item{name}{Name of the variable. A dimension with this name will also be created.}

\item{values}{Values for the variable.}

\item{type}{NetCDF type for \code{values}. If \code{NULL}, \code{type} is determined from \code{value}'s class. See
\link{NC_TYPES} for a list of valid types.}

\item{long_name}{Optional descriptive name of the variable.}

\item{standard_name}{Optional standard name of the variable.}

\item{axis}{Optional axis (X, Y, Z, or T) represented by the variable.}

\item{units}{Optional units of the variable.}

\item{atts}{Optional list of attributes created with \code{\link[=nc_att]{nc_att()}}.}

\item{decreasing}{Whether the values should be sorted in decreasing order.}

\item{unlim}{Whether the dimension should be unlimited.}

\item{fill}{Whether to use \code{\link[=fill_seq]{fill_seq()}} to ensure values differ by a constant increment. This is
useful for coordinates of truncated regular grids (i.e. a regular grid where the coordinates of
some points are omitted because they have no data). When writing such data to a netCDF file, the
coordinates must contain a value for \emph{all} points in the grid domain and the data variables must
have NA at coordinates with no data. Otherwise the data for the point preceding an omitted
coordinate will be interpreted as having an increased size along the omitted coordinate's
dimension (e.g. if one x-coordinate is omitted, the data for the previous x-coordinate will be
interpreted as having a width of two x-coordinates).}

\item{...}{Arguments passed to \code{\link[=fill_seq]{fill_seq()}}.}
}
\value{
A list of the passed arguments
}
\description{
Define a netCDF coordinate variable (dimension and variable)
}
\examples{
nc_coord("x", values = 1:5, type = "NC_INT")

nc_coord("x", values = c(1, 2, 3, 5), fill = TRUE, atts = nc_att("long_name", "x coordinate"))
}
\seealso{
See https://cfconventions.org/ for recommended names, axis, and units.
}
