% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{nc_crs}
\alias{nc_crs}
\title{Define a netCDF variable describing a coordinate reference system (CRS)}
\usage{
nc_crs(
  name,
  grid_mapping_name,
  crs_wkt = NULL,
  earth_radius = NULL,
  inverse_flattening = NULL,
  longitude_of_prime_meridian = NULL,
  prime_meridian_name = NULL,
  reference_ellipsoid_name = NULL,
  semi_major_axis = NULL,
  semi_minor_axis = NULL,
  atts = NULL
)
}
\arguments{
\item{name}{Name of the variable.}

\item{grid_mapping_name}{Value for the 'grid_mapping_name' attribute.}

\item{crs_wkt}{Optional well-known text representation of the CRS.}

\item{earth_radius}{Optional radius (m) of sphere used to approximate the earth.}

\item{inverse_flattening}{Optional inverse flattening of ellipsoid used to approximate the earth.}

\item{longitude_of_prime_meridian}{Optional long. of prime meridian with respect to Greenwich.}

\item{prime_meridian_name}{Optional name of prime meridian.}

\item{reference_ellipsoid_name}{Optional name of ellipsoid used to approximate the earth.}

\item{semi_major_axis}{Optional semi-major axis (m) of ellipsoid used to approximate the earth.}

\item{semi_minor_axis}{Optional semi-minor axis (m) of ellipsoid used to approximate the earth.}

\item{atts}{Optional list of grid mapping-specific attributes created with \code{\link[=nc_att]{nc_att()}}.}
}
\value{
A list of the passed arguments.
}
\description{
Define a netCDF variable describing a coordinate reference system (CRS)
}
\examples{
nc_crs("crs", grid_mapping_name = "sinusoidal", earth_radius = 6371007.181)
}
\seealso{
See https://cfconventions.org/ section on "Coordinate Systems and Domain" for guidance on
documenting coordinate systems.
}
