% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_date.R
\name{nc_date}
\alias{nc_date}
\title{nc_date}
\usage{
nc_date(fid, ntime = -1, to_char = FALSE)
}
\arguments{
\item{fid}{An object of class ncdf4 retuned by \code{\link[ncdf4:nc_open]{ncdf4::nc_open()}}}

\item{ntime}{How many time to read?}

\item{datastr}{Boolean. Whether convert date to string?}
}
\description{
nc_date
}
\keyword{internal}
