% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_format.R
\name{nc_format}
\alias{nc_format}
\title{Format the data frame into gridded matrix}
\usage{
nc_format(df, df.setdims)
}
\arguments{
\item{df}{dataframe containing all the data}

\item{df.setdims}{dataframe with the dimension informations}
}
\description{
Format the data frame into gridded matrix
}
