% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_rewrite.R
\name{nc_rewrite}
\alias{nc_rewrite}
\title{rewrite netcdf to fix attribute information}
\usage{
nc_rewrite(
  x,
  outfile,
  bands = NULL,
  var.units = NULL,
  var.longname = NULL,
  dates = NULL,
  info = NULL
)
}
\arguments{
\item{bands}{variable names}

\item{var.units}{untis of variables in \code{lst}}

\item{var.longname}{longname of variables in \code{lst}}

\item{dates}{A \code{Date} or \code{POSIXct} object, specified corresponding date}
}
\description{
rewrite netcdf to fix attribute information
}
\examples{
\dontrun{
nc_rewrite(file, outfile, bands, units, longnames, dates)
}
}
