% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_setdims.R
\name{nc_setdims}
\alias{nc_setdims}
\title{Inform on the dimensions for each variable}
\usage{
nc_setdims(df, dims)
}
\arguments{
\item{df}{input dataframe containing the gridded variables and dimensions}

\item{dims}{character array indicating the dimensions}
}
\description{
Inform on the dimensions for each variable
}
