% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{nc_var}
\alias{nc_var}
\title{Define a netCDF variable}
\usage{
nc_var(
  name,
  data,
  dims,
  type = NULL,
  pack_type = NULL,
  add_offset = NULL,
  scale_factor = NULL,
  long_name = NULL,
  standard_name = NULL,
  units = NULL,
  grid_mapping = NULL,
  fill_value = NULL,
  atts = NULL,
  chunking = NA,
  deflate = NA,
  shuffle = FALSE
)
}
\arguments{
\item{name}{Name of the variable.}

\item{data}{A \code{data.frame} or \code{data.table} containing the variable's coordinates and values. The
column names must exactly match \code{dims} (the coordinates) and \code{name} (the values).}

\item{dims}{The names of the variable's dimensions. The last dimension varies fastest.}

\item{type}{NetCDF type for the values. If \code{NULL}, \code{type} is determined from the values' class.
If values are to be packed (see below), \code{type} must be the type of the \emph{unpacked} (original)
values; the packed values will be written to the file as type \code{pack_type}. The resulting list
will have \code{type} set to \code{pack_type} to indicate that the data will be written as \code{pack_type}. See
\link{NC_TYPES} for a list of valid types.}

\item{pack_type, add_offset, scale_factor}{If defined, values will be packed to save space by first
subtracting \code{add_offset} then dividing by \code{scale_factor} and writing as type \code{pack_type}.}

\item{long_name}{Optional descriptive name of the variable.}

\item{standard_name}{Optional standard name of the variable.}

\item{units}{Optional units of the variable.}

\item{grid_mapping}{Optional name of a grid_mapping created with \code{\link[=nc_crs]{nc_crs()}}. For details, see}

\item{fill_value}{Optional value to use for missing data. If data are packed, \code{fill_value} must
be within range of For details, see}

\item{atts}{Optional list of attributes created with \code{\link[=nc_att]{nc_att()}}.}

\item{chunking}{Set to \code{NA} to allow the netCDF library to choose a storage layout, \code{FALSE} for
contiguous storage, or a vector named with the dimensions indicating the number of elements along
each dimension. Any \code{NA} elements in the vector will be replaced by the dimension's size.}

\item{deflate}{Compression level from 1 (min) to 9 (max) or \code{NA} (no compression).}

\item{shuffle}{Whether to enable byte shuffling, which may improve compression with \code{deflate}.}
}
\value{
A list of the passed arguments.
}
\description{
Define a netCDF variable
}
\examples{
## Generate some data
data <- cbind(expand.grid(x = 1:3, y = 1:3), list(elev = 0:8 * 12.5))

## Define a variable
nc_var("elev", data = data, dims = c("y", "x"), atts = nc_att("long_name", "elevation"))

## Chunk data with each chunk containing 1 y-coordinate and all x-coordinates
nc_var("elev", data = data, dims = c("y", "x"), chunking = c(y = 1, x = NA))

## Pack to reduce file size: divide by 0.1 and write as 16-bit integers rather than 32-bit floats
nc_var("elev", data = data, dims = c("y", "x"), pack_type = "NC_SHORT", scale_factor = 0.1)
}
\seealso{
See https://cfconventions.org/ for details on packing, grid mappings, and recommended
names and units. See https://docs.unidata.ucar.edu/nug/current/netcdf_perf_chunking.html for
details on chunking.
}
