% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\name{nc_write}
\alias{nc_write}
\title{Create a netCDF file}
\usage{
nc_write(
  path,
  coords,
  vars,
  crs = NULL,
  atts = NULL,
  force = FALSE,
  quiet = FALSE,
  log_level = logger::WARN,
  ...
)
}
\arguments{
\item{path}{Output path.}

\item{coords}{List of coordinate variables defined with \code{\link[=nc_coord]{nc_coord()}}.}

\item{vars}{List of variables defined with \code{\link[=nc_var]{nc_var()}}.}

\item{crs}{Optional coordinate reference system defined with \code{\link[=nc_crs]{nc_crs()}}.}

\item{atts}{Optional global attributes defined with \code{\link[=nc_att]{nc_att()}}.}

\item{force}{Whether to overwrite \code{path} if it exists.}

\item{quiet}{Whether to silence warnings about missing attributes suggested by CF conventions.}

\item{log_level}{Logging level for \link{logger} messages.}

\item{...}{Arguments passed to \code{\link[RNetCDF:var.put.nc]{RNetCDF::var.put.nc()}}}
}
\description{
Create a netCDF file
}
\examples{
## Write a variable to a netCDF file
#
}
