% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncatt_get_all}
\alias{ncatt_get_all}
\title{Get all attributes from variables or dimensions}
\usage{
ncatt_get_all(nc, type = c("var", "dim"))
}
\arguments{
\item{nc}{An open netCDF connection created with [ncdf4::nc_open()].}

\item{type}{Character string indicating whether to return attributes for
variables (`"var"`) or dimensions (`"dim"`).}
}
\value{
A named list of attribute lists.
}
\description{
Returns all attributes associated with variables or dimensions in an open
netCDF file.
}
\examples{
\dontrun{
nc <- ncdf4::nc_open("input.nc")
on.exit(ncdf4::nc_close(nc))

ncatt_get_all(nc, type = "var")
ncatt_get_all(nc, type = "dim")
}
}
\seealso{
[ncatt_put_all()], [ncdf4::ncatt_get()]
}
