% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncatt_put_all}
\alias{ncatt_put_all}
\title{Write multiple attributes to a netCDF variable or file}
\usage{
ncatt_put_all(
  nc,
  varid,
  attname,
  attval,
  prec = NA,
  verbose = FALSE,
  definemode = FALSE
)
}
\arguments{
\item{nc}{An open netCDF connection created with [ncdf4::nc_open()].}

\item{varid}{Variable identifier. This can be a character string giving the
variable name, an object of class `ncvar4`, or an integer id. As a special
case, `varid = 0` writes global attributes.}

\item{attname}{Names of the attributes to write. Alternatively, this may be a
named vector or list of attribute values when `attval` is omitted.}

\item{attval}{Values of the attributes to write. If omitted, `attname` must
be a named vector or list whose names are used as attribute names.}

\item{prec}{Optional precision used when writing the attributes. Passed to
[ncdf4::ncatt_put()].}

\item{verbose}{Logical. If `TRUE`, print additional information while writing
attributes.}

\item{definemode}{Logical. Passed to [ncdf4::ncatt_put()]. See that function
for details.}
}
\value{
`NULL`, invisibly.
}
\description{
Writes several attributes to a variable or to the global file metadata of an
open netCDF file.
}
\details{
Attributes can be supplied either as parallel `attname` and `attval`
arguments, or as a single named vector or list. Missing attribute names are
not allowed.
}
\examples{
\dontrun{
nc <- ncdf4::nc_open("input.nc", write = TRUE)
on.exit(ncdf4::nc_close(nc))

ncatt_put_all(
  nc,
  varid = "temp",
  attname = c("long_name", "units"),
  attval = list("Temperature", "degree_C")
)

ncatt_put_all(
  nc,
  varid = 0,
  attval = list(title = "Example file", source = "nctools")
)
}
}
\seealso{
[ncatt_get_all()], [ncdf4::ncatt_put()]
}
