% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdim_def.R
\name{ncdim_def}
\alias{ncdim_def}
\alias{ncdim_def_lonlat}
\alias{ncdim_def_range}
\title{Define a netCDF Dimension}
\usage{
ncdim_def_lonlat(lons, lats, dates = NULL)

ncdim_def_range(dim, range = c(-180, 180, -90, 90), dates = NULL)
}
\description{
Define a netCDF Dimension
}
