% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncdim_size}
\alias{ncdim_size}
\title{Get dimension lengths from a netCDF file}
\usage{
ncdim_size(nc)
}
\arguments{
\item{nc}{An open netCDF connection created with [ncdf4::nc_open()].}
}
\value{
A named list giving the length of each dimension.
}
\description{
Returns the lengths of all dimensions in an open netCDF file.
}
\examples{
\dontrun{
nc <- ncdf4::nc_open("input.nc")
on.exit(ncdf4::nc_close(nc))

ncdim_size(nc)
}
}
\seealso{
[ncvar_dim()], [ncdim_isUnlim()]
}
