% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncread.R
\name{ncread}
\alias{ncread}
\alias{ncread_all}
\title{ncread}
\usage{
ncread(file, varnames = 1L, period = NULL, ...)

ncread_all(file)
}
\arguments{
\item{file}{file path}

\item{varnames}{\itemize{
\item \code{-1}: all variables
}}

\item{...}{others to \code{\link[=ncvar_get]{ncvar_get()}}}
}
\description{
ncread
}
