% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-package.R
\docType{package}
\name{nctools-package}
\alias{nctools}
\alias{nctools-package}
\title{nctools: Tools tidy nc files}
\description{
Read, write, and process NETCDF files.
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
