% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools-auxiliar.R
\name{ncvar_change_compression}
\alias{ncvar_change_compression}
\title{Change variable compression settings in a netCDF object}
\usage{
ncvar_change_compression(nc, varid = NA, compression)
}
\arguments{
\item{nc}{An open netCDF connection created with [ncdf4::nc_open()].}

\item{varid}{Variable identifier. This can be a character string giving the
variable name, or an object of class `ncvar4`. If `NA`, the compression
setting is changed for all variables in the file.}

\item{compression}{Numeric compression level to assign.}
}
\value{
The modified netCDF object.
}
\description{
Modifies the compression setting of one or more variables in an open netCDF
object.
}
\examples{
\dontrun{
nc <- ncdf4::nc_open("input.nc")
on.exit(ncdf4::nc_close(nc))

nc <- ncvar_change_compression(nc, varid = "temp", compression = 4)
}
}
\seealso{
[ncvar_compression()]
}
