% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncwrite.R, R/ncwrite_raster2.R
\name{ncwrite}
\alias{ncwrite}
\alias{ncwrite_raster2}
\title{Write CMIP5 nc file}
\usage{
ncwrite(
  lst,
  file,
  var.units = NULL,
  var.longname = NULL,
  prec = "float",
  dims = NULL,
  dimnames_last = NULL,
  range = c(-180, 180, -90, 90),
  dates = NULL,
  calendar = "gregorian",
  missval = -9999L,
  compression = 1,
  scale,
  offset,
  attrs,
  overwrite = FALSE,
  verbose = FALSE
)

ncwrite_raster2(r, file, dims = NULL, dimnames_last, overwrite = TRUE, ...)
}
\arguments{
\item{lst}{A named list of 3d arrays, which have the dimension of \verb{[nlon, nlat, ntime]}.
\code{image(lst[[1]][,,1])} should looks normal.}

\item{file}{Name of the netCDF file to be created.}

\item{var.units}{untis of variables in \code{lst}}

\item{var.longname}{longname of variables in \code{lst}}

\item{prec}{Precision of the created variable. Valid options: 'short' 'integer'
'float' 'double' 'char' 'byte'. See the special note below for how to create
a character variable (strings).
short: -32768 - 32768.}

\item{range}{A numeric vector, \verb{[lon_min, lon_max, lat_min, lat_max]}.}

\item{dates}{A \code{Date} or \code{POSIXct} object, specified corresponding date}

\item{missval}{The variable's missing value. If NO missing value is desired, pass a NULL, or omit this argument entirely. If a NaN missing value is desired, pass an NA.}

\item{compression}{If set to an integer between 1 (least compression) and 9 (most compression), this
 enables compression for the variable as it is written to the file.
 Turning compression on forces the created file to be in netcdf version 4 format,
 which will not be compatible with older software that only reads netcdf version 3 files.}

\item{scale, offset}{If the variable in the netCDF file has a scale and/or offset
attribute defined, the returned data are automatically and silently scaled
and/or offset as requested. Then you should just leave scale = 1,
and offset = 0.}

\item{attrs}{A named list of global attributes, e.g. title, author, souce,
references, institution.}

\item{verbose}{Print debugging information.}

\item{r}{\code{raster2} object returned by \code{\link[=ncwrite]{ncwrite()}}}

\item{...}{other parameters to \link{ncwrite}}

\item{varname}{variable name in the data of \code{r}}
}
\description{
Write CMIP5 nc file
}
\examples{
\dontrun{
range_china = c(70, 140, 15, 55)
ncwrite(list(Tmin = arr[,,1:2]), "Tmin.nc", "degC", range = range_china,
     prec = "short", scale = 0.01, offset = 10,
     dates = dates)
}

}
\seealso{
\code{\link[=ncread_cmip]{ncread_cmip()}}, \code{\link[ncdf4:ncvar_def]{ncdf4::ncvar_def()}}
}
