% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2_plot.R
\name{plot.raster2}
\alias{plot.raster2}
\title{plot raster2 object}
\usage{
\method{plot}{raster2}(
  x,
  varname,
  zcols,
  I_sel,
  range = NULL,
  fix_lon360 = TRUE,
  offset = 0,
  sp.layout = NULL,
  is.SpatialMat = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{cmip5} object returned from ncread}

\item{varname}{variable name in the data of \code{x}}

\item{zcols}{how many columns to plot? default is 4.}

\item{I_sel}{A vector indicating subset rows of \code{x}}

\item{fix_lon360}{boolean. If true, it will convert longitude from 0-360 to
-180 - 180.}

\item{is.SpatialMat}{is it spatial matrix (FALSE: numeric vector)?}

\item{...}{other parameters to \code{spplot}}
}
\value{
\itemize{
\item \code{plot}: \code{spplot} plot ojbect
\item \code{grid}: SpatialGridDataframe
}
}
\description{
plot raster2 object
}
