% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_convert.R
\name{tiff2nc_merge}
\alias{tiff2nc_merge}
\alias{tiff2nc_process}
\title{convert tif into nc files and merge}
\usage{
tiff2nc_merge(files, outfile = "merged.nc", overwrite = FALSE, cache = TRUE)

tiff2nc_process(
  files,
  outfile,
  var.units = NULL,
  var.longname = NULL,
  run = FALSE
)
}
\arguments{
\item{files}{tiff file paths}

\item{outfile}{output nc file}

\item{overwrite}{If \code{overwrite}, temporal nc files and \code{outfile} will be overwrited.}

\item{cache}{If \code{cache} = false, temporal nc files will be removed.}

\item{var.units}{untis of variables in \code{lst}}

\item{var.longname}{longname of variables in \code{lst}}

\item{verbose}{boolean}
}
\description{
convert tif into nc files and merge by \code{gdal_translate} and \code{cdo}
}
\note{
This function not works for windows system.
}
\examples{
\dontrun{
tiff2nc_process(files, outfile, var.units, var.longname, run = FALSE)
}

}
