% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_merge.R
\name{nc_merge}
\alias{nc_merge}
\title{Read data from cmip5 nc files and merge into a large array}
\usage{
nc_merge(
  d,
  varname = 1L,
  range,
  delta = 2,
  verbose = TRUE,
  bigmemory = FALSE,
  scale = 1,
  offset = 0,
  value_range = NULL,
  adjust_lon = TRUE
)
}
\arguments{
\item{d}{A data.frame at least with the columns of 'start_adj', 'end_adj' and 'file'`.}

\item{range}{A numeric vector, \verb{[lon_min, lon_max, lat_min, lat_max]}. If \code{range}
is \code{NULL}, \code{delta} will be ignored, and all data will be read.}

\item{delta}{suggest 3deg. Extend range into \verb{[lat_min - delta, lat_max + delta, lon_min - delta, lon_max + delta]}. This is to make sure no missing value in
margin.}

\item{verbose}{whether to show reading progress?}

\item{bigmemory}{Boolean. If true, 3d array memory will be pre-allocated.
To do this, we need to open nc files first than get the dimension information.
This step will sacrifice some time. If false, temporary data are stored in
list and finally convert to 3d array.}

\item{scale}{If the variable in the netCDF file has a scale and/or offset
attribute defined, the returned data are automatically and silently scaled
and/or offset as requested. Then you should just leave scale = 1,
and offset = 0.}

\item{offset}{If the variable in the netCDF file has a scale and/or offset
attribute defined, the returned data are automatically and silently scaled
and/or offset as requested. Then you should just leave scale = 1,
and offset = 0.}

\item{value_range}{If not null, values out of \code{value_range} are set to
NA.}

\item{adjust_lon}{If true, lon = lon + cellsize_y/2}
}
\value{
\itemize{
\item \code{grid.origin}: A list with \verb{'lon', 'lat', 'cellsize_x', 'cellsize_y'}
of original data.
\item \code{grid}:  A list with \verb{'lon', 'lat', 'date'} of the read data.
\item \code{data}: Merged array value, with the dimension of \verb{[nlon*nlat, ntime]}.
}
}
\description{
Read data from cmip5 nc files and merge into a large array
}
\note{
Only merge a model for each time. Please make sure that no duplicated date
in \code{d}.
}
\examples{
\dontrun{
library(magrittr)
range <- c(15, 55, 70,140)

d_files   <- CMIP5Files_filter(files)
lst_dfiles <- d_files \%>\% split(., .$model)

i = 1
d <- lst_dfiles[[i]]

l <- nc_merge(d, varname = 'tasmax', range, delta = 2)
}
}
