% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc_write.R
\name{nc_write}
\alias{nc_write}
\title{Write the gridded netcdf file}
\usage{
nc_write(fn, l)
}
\arguments{
\item{fn}{name of the output file}

\item{l}{list of the variable and dimension informations}
}
\description{
Write the gridded netcdf file
}
