% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncdim_def.R
\name{ncdim_def}
\alias{ncdim_def}
\alias{ncdim_def_lonlat}
\alias{ncdim_def_range}
\title{Define a netCDF Dimension}
\usage{
ncdim_def_lonlat(lons, lats, dates = NULL, ...)

ncdim_def_range(
  dim,
  range = c(-180, 180, -90, 90),
  dates = NULL,
  calendar = "gregorian"
)
}
\arguments{
\item{dates}{PCICt object. The default origin is '1970-01-01'}

\item{...}{ignored parameters}
}
\description{
Define a netCDF Dimension
}
