% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncread.R
\name{ncread}
\alias{ncread}
\title{read nc file}
\usage{
ncread(
  file,
  varnames = 1L,
  range = NULL,
  delta = 0,
  ntime = -1,
  DatePeriod,
  grid_type = NULL,
  scale = 1,
  offset = 0,
  value_range = NULL,
  adjust_lon = FALSE,
  check_date = FALSE
)
}
\arguments{
\item{file}{Character, nc file name.}

\item{varnames}{String or integer vector which specify which variable to read.
if integer vector, for example: \code{1} means the first variable to read;
\itemize{
\item \code{0} or \code{NULL}: no data to be read, just return dimension infomation
\item \code{-1}         : read all data
}}

\item{range}{A numeric vector, \verb{[lon_min, lon_max, lat_min, lat_max]}. If \code{range}
is \code{NULL}, \code{delta} will be ignored, and all data will be read.}

\item{delta}{suggest 3deg. Extend range into \verb{[lat_min - delta, lat_max + delta, lon_min - delta, lon_max + delta]}. This is to make sure no missing value in
margin.}

\item{ntime}{How many time to read?}

\item{DatePeriod}{Date \strong{character} vector, starting and ending date.
\code{DatePeriod} works only in the situation when \code{time2date} is true.
Note that in that situation if \code{DatePeriod} provided, ntime will be overrided.}

\item{grid_type}{'vec', 'mat' or NULL. If not null, 3d array will be converted
to 2d matrix, with the dimension of \verb{[nlon*nlat, ntime]}.
\itemize{
\item \code{vec}: spdata_array(vals, flip = TRUE) will be applied.
\item \code{mat}: spdata_array(vals, flip = FALSE)
}}

\item{scale, offset}{If the variable in the netCDF file has a scale and/or offset
attribute defined, the returned data are automatically and silently scaled
and/or offset as requested. Then you should just leave scale = 1,
and offset = 0.}

\item{value_range}{If not null, values out of \code{value_range} are set to
NA.}

\item{adjust_lon}{If true, lon = lon + cellsize_y/2}

\item{check_date}{If true, check whether date in nc file is consistent with
\code{date_begin} and \code{date_end} in its filename.}
}
\value{
list(dim, date, value)
}
\description{
\code{getCMIP5date} make sure that if dates geenrated according to
\code{time} and dates in files consistent. If not, error will occur.
}
\note{
In some CMIP5 file, \code{time} band is incorrect. So we decide to generate
date by the date information containing in the file name.

This function is initially designed for CMIP5 nc files.
}
\examples{
\dontrun{
file <- "H:/CMIP5/historical_tasmax/tasmax_day_ACCESS1-0_historical_r1i1p1_20000101-20051231.nc"

range <- c(0, 360, -90, 90)
range <- c(70,140, 15, 55)
l <- ncread(file, varnames = "tasmax", range = range, delta = 2, ntime = 1)
}

}
\seealso{
spdata_array
}
