% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nctools.R
\docType{package}
\name{nctools-package}
\alias{nctools}
\alias{nctools-package}
\title{Some tools for creating and playing around with netcdf data
TBC}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Odran Sourdeval \email{odran.sourdeval@univ-lille.fr}

}
