% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdata_array.R
\name{spdata_array}
\alias{spdata_array}
\title{Convert Spatial* data into 3d array}
\usage{
spdata_array(vals, range = NULL, cellsize = 0.1, flip = TRUE, to2d = FALSE)
}
\arguments{
\item{vals}{\verb{[ngrid, ntime]}
If ndim <= 2, range and cellsize is necessary.}
}
\description{
Convert Spatial* data into 3d array
}
\keyword{internal}
